hackon_tts/ai call/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# Audio Processing
soundfile>=0.12.1
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0

# Speech-to-Text
faster-whisper>=0.10.0
onnxruntime>=1.16.0

# Text-to-Speech
elevenlabs>=1.0.0

# LLM/AI Services
google-genai>=0.7.0

# WebSocket Support
websockets>=12.0

# File Upload Support
python-multipart>=0.0.6

# Optional: ML/Transformers (commented out in current code but available)
torch>=2.1.0
# transformers>=4.35.0
# peft>=0.7.0
